Update relevant higher level classes with module add task methods
